Add basic structure and configuration
Add basic structure for started new projects with less effort.

In the basic configuration we only assume the next sections: testing,
linter, github actions and styles. Because we think that these are the
most important for start a new project.

Also we add a basic structure for projects in Next.js. This structure
is opinionated.